+++ src/Model/Table/PagesTable.php
@@ -80,6 +80,8 @@ class PagesTable extends AppTable
     public function buildRules(RulesChecker $rules)
     {
         $rules->add($rules->existsIn(['category_id'], 'Categories'));
+        $rules->add($rules->isUnique(['slug'], __d('me_cms', 'This value is already used')));
+        $rules->add($rules->isUnique(['title'], __d('me_cms', 'This value is already used')));
 
         return $rules;
     }