+++ src/Model/Table/PhotosAlbumsTable.php
@@ -24,6 +24,7 @@ namespace MeCms\Model\Table;
 
 use Cake\Filesystem\Folder;
 use Cake\ORM\Query;
+use Cake\ORM\RulesChecker;
 use MeCms\Model\Table\AppTable;
 
 /**
@@ -71,6 +72,20 @@ class PhotosAlbumsTable extends AppTable
         }
 
         parent::afterSave($event, $entity, $options);
+    }
+
+    /**
+     * Returns a rules checker object that will be used for validating
+     *  application integrity
+     * @param \Cake\ORM\RulesChecker $rules The rules object to be modified
+     * @return \Cake\ORM\RulesChecker
+     */
+    public function buildRules(RulesChecker $rules)
+    {
+        $rules->add($rules->isUnique(['slug'], __d('me_cms', 'This value is already used')));
+        $rules->add($rules->isUnique(['title'], __d('me_cms', 'This value is already used')));
+
+        return $rules;
     }
 
     /**