+++ src/Model/Table/PostsCategoriesTable.php
@@ -48,6 +48,8 @@ class PostsCategoriesTable extends AppTable
     public function buildRules(RulesChecker $rules)
     {
         $rules->add($rules->existsIn(['parent_id'], 'Parents'));
+        $rules->add($rules->isUnique(['slug'], __d('me_cms', 'This value is already used')));
+        $rules->add($rules->isUnique(['title'], __d('me_cms', 'This value is already used')));
 
         return $rules;
     }