+++ src/Model/Table/TagsTable.php
@@ -23,6 +23,7 @@
 namespace MeCms\Model\Table;
 
 use Cake\ORM\Query;
+use Cake\ORM\RulesChecker;
 use Cake\Validation\Validator;
 use MeCms\Model\Table\AppTable;
 
@@ -37,6 +38,19 @@ class TagsTable extends AppTable
      * @var string
      */
     public $cache = 'posts';
+
+    /**
+     * Returns a rules checker object that will be used for validating
+     *  application integrity
+     * @param \Cake\ORM\RulesChecker $rules The rules object to be modified
+     * @return \Cake\ORM\RulesChecker
+     */
+    public function buildRules(RulesChecker $rules)
+    {
+        $rules->add($rules->isUnique(['tag'], __d('me_cms', 'This value is already used')));
+
+        return $rules;
+    }
 
     /**
      * "Active" find method