+++ tests/TestCase/Model/Validation/PagesCategoryValidatorTest.php
@@ -86,4 +86,36 @@ class PagesCategoryValidatorTest extends TestCase
             $this->assertEquals([$key => ['_required' => 'This field is required']], $errors);
         }
     }
+
+    /**
+     * Test validation for `slug` property, testing that is unique
+     * @test
+     */
+    public function testValidationForSlugIsUnique()
+    {
+        $entity = $this->PagesCategories->newEntity($this->example);
+        $this->assertNotEmpty($this->PagesCategories->save($entity));
+
+        //Saves again the same entity
+        $this->example['title'] = 'New title';
+        $entity = $this->PagesCategories->newEntity($this->example);
+        $this->assertFalse($this->PagesCategories->save($entity));
+        $this->assertEquals(['slug' => ['_isUnique' => 'This value is already used']], $entity->errors());
+    }
+
+    /**
+     * Test validation for `title` property, testing that is unique
+     * @test
+     */
+    public function testValidationForTitleIsUnique()
+    {
+        $entity = $this->PagesCategories->newEntity($this->example);
+        $this->assertNotEmpty($this->PagesCategories->save($entity));
+
+        //Saves again the same entity
+        $this->example['slug'] = 'new-slug';
+        $entity = $this->PagesCategories->newEntity($this->example);
+        $this->assertFalse($this->PagesCategories->save($entity));
+        $this->assertEquals(['title' => ['_isUnique' => 'This value is already used']], $entity->errors());
+    }
 }